1. Реализован алгоритм ипорта файлов оснастки на основе идентификатора "ID_DOC". 2. Доработано окно "О программе". 3. Инициализировано стандартное меню для элементов управления. 4. Добавлена задача связи технологических операций для ипорта ТП.

+++ omp_sepo/views/OsnTypesListView.cs
@@ -50,7 +50,7 @@ namespace omp_sepo.views
         private void OsnTypesListView_MouseDoubleClick(object sender, MouseEventArgs e)
         {
             ListViewItem row = this.SelectedItems[0];
-            decimal id = (decimal)row.Tag;
+            int id = (int)row.Tag;
 
             OsnTypesDialog dialog = new OsnTypesDialog(id);
             if (dialog.ShowDialog() == DialogResult.OK)
